backend/app/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-barcode
Pillow
